requirements.txt
influxdb-4.0.0
flask-0.12
gunicorn-19.6.0